Move services into own folder and namespace
