+++ backend/FwHeadless/Program.cs
@@ -1,4 +1,5 @@
 using FwHeadless;
+using FwHeadless.Services;
 using FwDataMiniLcmBridge;
 using FwDataMiniLcmBridge.Api;
 using FwLiteProjectSync;
